Massive language, library, and misc changes.
spectre language changes:
- C/C++ style comments were added to the language. More specifically, '//' can
be used for single line comments and '/*' and '*/' can delimit multi-line
comments.
- An 'auto' keyword was added for type deduction. Now, C++-style auto statements
like 'auto x = 12;' can be written, and 'x' will be deduced to an 'int'.
- A 'constexpr' keyword was added to the language. Now, 'constexpr' variables
can be created in spectre. The main reason for this is a type-safe alternative
to C/C++ '#define' preprocessor directives. 'constexpr' variables can be used in
any context an integral constant or pointer can be used.
- A 'delete' keyword was added to complement the 'new' keyword. Unlike C++,
variables allocated with 'new' or 'malloc' can be freed with 'free' or 'delete'.
- Fixed a bug which prevented namespace aliases from being used in function and
block scopes.

libspectre changes:
- Added 'std/aliases' and 'stdx/aliases' headers to simplify the usage of
members in 'std::' and 'stdx::' namespaces.
- Changed all the constants in the headers and implementation files when
possible to be 'constexpr'.
- Fixed a bug in 'std::lib::free'.
- Added a 'stdx/algorithm' and 'stdx/iterator' to allow for
container-independent algorithms in the future. The appropriate changes to
support this functionality were added to 'stdx/vector'.
- Added some more functions to 'stdx/string'.
- Added an 'stdx/allocator' header to potentially model a container-independent
allocator which doesn't have to rely on 'malloc'/'free'/'realloc' (although it
can be configured to do so).

misc changes:
- Added more tests.
- Added an automated way to check outputs and a set of reference outputs.
- Cleaned up and added the following new samples: an expression evaluator
(supports C expressions while following operator precedence), refactored out an
ltoa function, a text-spacing program, a simple clone of the 'cat' UNIX program,
a simple clone of the 'true', 'false', and 'yes' programs, a program which
mimics some of the functionality of the 'ls -la', and a string sorting progam
(mimics some of the functionality of the UNIX program of 'sort') which reads and
sorts strings from stdin.