Massive language and library changes.
This will be a partial list of changes itemized by which portion of the
language/compiler/library the change falls into. Since this commit is
relatively big, this list of changes might be incomplete.
1. Language
- Added support for variadic functions in the language. So variadic
  functions can now be written in Spectre and there is support for
variadic functions in the type system as well.
- Statement expressions are also now supported in the language as well
  (this is a GCC extension for the C language as well and the behavior
is pretty similar).
- Added support for unions in the language.

2. Library
- Added support for a heap based alloca (and thus, "freea"). Perhaps in
  the future alloca can be a builtin that can actually use the stack,
but because the compiler generates fixed-length frames for functions,
this is not possible now.
- Added support for a C-like va_arg and related functions to the
  standard library. This also enabled the implementation of a very
primitive printf function.
- Added some basic limits similar to limits.h in C.

3. Misc runtime
- Added a script to automate the building and cleaning of samples.
- Added some more tests, most of which exercise the new features added
  to the language as noted above.

4. Compiler
This is perhaps the biggest change in this commit (in terms of LOC as
well as functionality). This compiler for Spectre has begun to
transition away from an AST to assembly codegen backend to an AST to IR
(and eventually IR to assembly) codegen backend. With this is in mind,
here is what this commit does to achieve this goal:
- Adds code to generate an IR that supports (or rather should support),
  targetting the entire Spectre language as of this commit.
- Adds code to split the generated IR into basic blocks and compute a
  CFG where nodes are the basic blocks.
- Adds a framework for running optimization passes (this is probably
  still a WIP and the interface will most likely change quite a bit
before stabilizing).
- Adds support for the following optimizations:
> Converts memory locations that only store and load (and that aren't
structs), into registers.
> Supports conversion to SSA from regular IR code. This includes a
dominance frontier-based approach to phi-node insertion. This has been
tested a bit, but I expect more fixes to come.
> Supports intra-basic block constant folding. This code heavily relies
on variants so compilation for that file is pretty slow and there is a
large blow-up in executable size due to a large number of std::visit
calls on multiple variants at once (this is for debug code), -Os and ICF
could potentially help here (at least with Visual Studio, code size is a
lot better than the current Makefile).
> Supports a basic form of DCE which just removes trivially unused
registers. This also leaves a over-complicated CFG with unreachable
basic blocks, so some simplification should be done after this.
Note, that the IR generating backend is now the default backend being
used, but the old AST to MIPS assembly backend still exists and can
these two backends can be toggled based on the corresponding flag in
config.hpp.