Fixed a stack alignment bug.
Corrected the backend ASM generation code which aligned stack variables
in a function call. Also fixed the code laying out stack parameters for
variadic functions; now this matches what the family of std/arg
functions expect.
To test this, I added support for the '%c' format specifier for the
printf function and included a test as well.