Optimized the register allocation algorithm.
Specifically, the register spilling process was optimized. Originally,
the algorithm identified a register, spilled it, and re-colored the
entire graph (with whatever setup needed to be done). Then the entire
process was repeated again and if a new register was identified, it was
spilled and the entire thing was re-done.
Now, all registers that are going to spilled are identified at once, and
are spilled one-by-one (without recomputing the entire interference
graph every time). This leads to a pretty large speed up.