Changed how the standard library was built.
The old process for building the standard library was through a build.sh
script. This has now been changed to a Makefile; this was mainly to ease
development, since now only files that have been updated have to be
built.
Note, this currently includes only source files (*.sp implementation
files). If a header (like std/math for example) changes, then this
won't be detected.