Added an optimization for pure function calls.
If there is a pure function call starting a postfix expression (like
m() for example), the IR no longer generates a temporary register, so
the assembly generate is a bit more efficient; instead of:

mov r12, _m
...
blx r12

we, can now generate:
...
bl _m

for example. So in other words, we can use a bl instead of a combination
of a mov and a blx for pure function calls starting a postfix
expression.
Added some more tests with function calls as well.