Fixed compiler signed-ness bug.
'char' is by default signed on x86, but is unsigned on ARM. As a result,
not specifying signed-ness in opt/fold_constants.{c,h}pp led to some
inconsistent behavior when running the compiler on machines between the
two architectures. Explicitly specified the sign while casting there to
avoid this.