Fixed codegen bug for extending to same bitwidth.
Previously, the ARM codegen backend elided extensions (zero or
signed) when the source and destination bitwidths were identical.
However, this is incorrect, since the operation is technically a move (not a
NOP). This caused failures whenever there was a "conversion" from an
into to a long (it would probably have caused failures in other places
as well).

This bug fix also fixed some bogus outputs in simple_ls as well.