More bugfixes.
- Fixed a subtle bug in the control flow graph simplification.
Essentially, every function should have at least one instruction
(corresponding to its return value location), in its header. However, if
the tail of a function is not reachable from a the beginning of the
function, this would completely optimize out the return value and
potentially clear the function header. This was a problem since the
header label could be removed resulting in weird behavior. This has been
fixed.
- For debug outputs, commented pseudo-instructions in the output now
follow the correct format for ARM comments (changed ';' to '@').